+++ src/app/modules/inventory/grid/actions.component.html
@@ -1,0 +1,9 @@
+<mat-form-field class="thin-field mr-2 w-60" appearance="outline">
+  <mat-label>Actions</mat-label>
+  <mat-select #actionSelect (selectionChange)="onAction($event.value.action, actionSelect)"
+    class="inventory-actions-dropdown text-sm">
+    @for (item of actions; track $index) {
+    <mat-option [value]="item" class="!h-3 text-sm" [disabled]="item.disabled">{{ item.name }}</mat-option>
+    }
+  </mat-select>
+</mat-form-field>
