+++ src/app/modules/inventory/inventory.component.ts
@@ -2,11 +2,11 @@ import { CommonModule } from '@angular/common'
 import type { OnDestroy, OnInit } from '@angular/core'
 import { Component, inject, ViewEncapsulation } from '@angular/core'
 import { MatButtonModule } from '@angular/material/button'
-import { MatDialog, MatDialogModule } from '@angular/material/dialog'
+import { MatDialogModule } from '@angular/material/dialog'
 import { MatExpansionModule } from '@angular/material/expansion'
 import { MatFormFieldModule } from '@angular/material/form-field'
 import { MatIconModule } from '@angular/material/icon'
-import { type MatSelect, MatSelectModule } from '@angular/material/select'
+import { MatSelectModule } from '@angular/material/select'
 import { MatTabsModule } from '@angular/material/tabs'
 import { MatTooltipModule } from '@angular/material/tooltip'
 import { ActivatedRoute, Router } from '@angular/router'
@@ -23,15 +23,15 @@ import { OrganizationService } from '@seed/api/organization'
 import { InventoryTabComponent, PageComponent } from '@seed/components'
 import { SharedImports } from '@seed/directives'
 import { InventoryGridComponent, InventoryGridControlsComponent } from './grid'
+import { ActionsComponent } from './grid/actions.component'
 import type { AgFilterResponse, FiltersSorts, InventoryPagination, InventoryType, Profile } from './inventory.types'
-import { DeleteModalComponent, MoreActionsModalComponent } from './modal'
-import { PopulatedColumnsModalComponent } from './modal/populated-columns-modal.component'
 
 @Component({
   selector: 'seed-inventory',
   templateUrl: './inventory.component.html',
   encapsulation: ViewEncapsulation.None,
   imports: [
+    ActionsComponent,
     CommonModule,
     MatButtonModule,
     MatDialogModule,
@@ -56,8 +56,6 @@ export class InventoryComponent implements OnDestroy, OnInit {
   private _organizationService = inject(OrganizationService)
   private _columnService = inject(ColumnService)
   private _labelService = inject(LabelService)
-  private _orgId: number = null
-  private _dialog = inject(MatDialog)
   private readonly _unsubscribeAll$ = new Subject<void>()
   readonly tabs: InventoryType[] = ['properties', 'taxlots']
   readonly type = this._activatedRoute.snapshot.paramMap.get('type') as InventoryType
@@ -69,12 +67,13 @@ export class InventoryComponent implements OnDestroy, OnInit {
   gridApi: GridApi
   labelLookup: Record<number, Label> = {}
   filters: string[][] | [] = []
+  inventory: Record<string, unknown>[]
+  orgId: number = null
   page = 1
   pagination: InventoryPagination
   profile: Profile
   profileId: number
   allProfiles: Profile[]
-  properties: Record<string, unknown>[]
   propertyColumns: Column[]
   propertyProfiles: Profile[]
   taxlotProfiles: Profile[]
@@ -104,10 +103,10 @@ export class InventoryComponent implements OnDestroy, OnInit {
   * get cycles, profiles, columns, inventory
   */
   getDependencies(org_id: number) {
-    this._orgId = org_id
+    this.orgId = org_id
 
     return forkJoin({
-      cycles: this._cycleService.get(this._orgId),
+      cycles: this._cycleService.get(this.orgId),
       profiles: this._inventoryService.getColumnListProfiles('List View Profile', 'properties', true),
       propertyColumns: this._columnService.propertyColumns$.pipe(take(1)),
       labels: this._labelService.labels$.pipe(take(1)),
@@ -178,7 +177,7 @@ export class InventoryComponent implements OnDestroy, OnInit {
       cycle: this.cycleId.toString(),
       ids_only: 'false',
       include_related: 'true',
-      organization_id: this._orgId.toString(),
+      organization_id: this.orgId.toString(),
       page: this.page.toString(),
       per_page: this.chunk.toString(),
     })
@@ -197,7 +196,7 @@ export class InventoryComponent implements OnDestroy, OnInit {
     }
     this._inventoryService.getAgInventory(this.type, paramString, data).subscribe(({ pagination, results, column_defs }: AgFilterResponse) => {
       this.pagination = pagination
-      this.properties = results
+      this.inventory = results
       this.columnDefs = column_defs
       this.rowData = results
     })
@@ -226,91 +225,6 @@ export class InventoryComponent implements OnDestroy, OnInit {
   openFilterSortModal() {
     console.log('open filter sort label')
   }
-
-  // SHOULD ACTIONS BE ITS OWN COMPONENT?
-  get actions() {
-    return [
-      { name: 'Select All', action: () => { this.selectAll() }, disabled: false },
-      { name: 'Select None', action: () => { this.deselectAll() }, disabled: false },
-      { name: 'Only Show Populated Columns', action: () => { this.openShowPopulatedColumnsModal() }, disabled: !this.properties },
-      { name: 'Delete', action: this.deletePropertyStates, disabled: !this.selectedViewIds.length },
-      { name: 'Merge', action: this.tempAction, disabled: !this.selectedViewIds.length },
-      { name: 'More...', action: () => { this.openMoreActionsModal() }, disabled: !this.selectedViewIds.length },
-    ]
-  }
-
-  tempAction() {
-    console.log('temp action')
-  }
-
-  openMoreActionsModal() {
-    this._dialog.open(MoreActionsModalComponent, {
-      width: '40rem',
-      autoFocus: false,
-      data: { viewIds: this.selectedViewIds, orgId: this._orgId },
-    })
-  }
-
-  onAction(action: () => void, select: MatSelect) {
-    action()
-    select.value = null
-  }
-
-  selectAll() {
-    this.gridApi.selectAll()
-    const params = new URLSearchParams({
-      cycle: this.cycleId.toString(),
-      ids_only: 'true',
-      include_related: 'true',
-      organization_id: this._orgId.toString(),
-    })
-    const paramString = params.toString()
-    this._inventoryService.getAgInventory(this.type, paramString, {}).subscribe(({ results }: { results: number[] }) => {
-      this.selectedViewIds = results
-    })
-  }
-
-  deselectAll() {
-    this.gridApi.deselectAll()
-  }
-
-  deletePropertyStates = () => {
-    const dialogRef = this._dialog.open(DeleteModalComponent, {
-      width: '40rem',
-      data: { orgId: this._orgId, viewIds: this.selectedViewIds },
-    })
-
-    dialogRef
-      .afterClosed()
-      .pipe(
-        takeUntil(this._unsubscribeAll$),
-        tap(() => { this.loadInventory() }),
-      )
-      .subscribe()
-  }
-
-  // openShowPopulatedColumnsModal() {
-  //   // if (!this.profiles.length) {
-  //   //   this.newProfile().subscribe(() => this.openPopulatedColumnsModal());
-  //   // } else {
-  //   //   this.openPopulatedColumnsModal();
-  //   // }
-  // }
-
-  openShowPopulatedColumnsModal() {
-    this._dialog.open(PopulatedColumnsModalComponent, {
-      width: '40rem',
-      data: {
-        orgId: this._orgId,
-        columns: null,
-        profile: this.profile,
-        cycleId: this.cycleId,
-        inventoryType: this.type,
-      },
-    })
-  }
-
-  // ^^ ACTIONS
 
   ngOnDestroy(): void {
     this._unsubscribeAll$.next()
